+++ include/Makefile.am
@@ -11,8 +11,8 @@ nobase_include_HEADERS = \
     fty_common_mlm_tntmlm.h \
     fty_common_mlm_utils.h \
     fty_common_mlm_zconfig.h \
-    fty-common-mlm-sync-client.h \
-    fty-common-mlm-stream-client.h \
+    fty_common_mlm_sync_client.h \
+    fty_common_mlm_stream_client.h \
     fty_common_mlm_library.h
