+++ src/Makemodule.am
@@ -19,8 +19,8 @@ src_libfty_common_mlm_la_SOURCES = \
     src/fty_common_mlm_tntmlm.cc \
     src/fty_common_mlm_utils.cc \
     src/fty_common_mlm_zconfig.cc \
-    src/fty-common-mlm-sync-client.cc \
-    src/fty-common-mlm-stream-client.cc \
+    src/fty_common_mlm_sync_client.cc \
+    src/fty_common_mlm_stream_client.cc \
     src/platform.h
 
 if ENABLE_DRAFTS
